Alpha 12
Alpha 12: The Job Update Pt. II
===============
Section 0: Introdution
===============
You can ignore this section. It contains no changelog information.
===============
This update is by far the biggest so far, more so than the previous one. I am really proud of the mechanics introduced, and cannot wait for more to come. In fact, this update has added so much, that I barely know what to add now. 

There is a [nearly] working savegame [with a new .OLL2 file extension] that allows you to load 99% of your character into the game, with a few things either set or randomised due to me forgetting to add them to the save file. Sadly, the old .OLL files are obsolete, and may not be back for a while. I will add a conversion utility eventually, but a lot of variables would have to be randomised. It isn't hard to do, it's just that I don't want to add it just yet. It's only an alpha game, after all.

When new updates come out, backwards compatibility will be a top priority. However, for obvious reasons, forwards compatibility will be impossible. Saves will be able to be converted to the latest version and launched. Naturally, a lot of data will be missing and will have to get added, but it should be doable, unless I rewrite the whole game from scratch or something. 

This update was originally going to be called "The Teen Update Part II", but I quickly realised how little BitLife has in terms of the teen stage anyway, so I didn't have much to add.

This update took forever to make, and I had so much to do, but I did it. I like how it turned out, and this game feels more and more fleshed out the more I add. 

Alpha 13 will be the Childbirth Update. It will allow you to:
[-] Find love
[-] Make love
[-] Marry
[-] Have up to 8 children
   [-] You will be able to have more in future. 
===============
Section 1: Teen Stage
[-] Added teen shop
   [-] Sells basic commodities
   [-] No items that you can get permanently
[-] Added Fake ID Hub
   [-] Allows you to use your fake ID to do things you               normally would not be able to.
===============
Section 2: Lottery Tweaks
We have made a few "tweaks" to the lottery system to make it more fair.
[-] Odds changed from 1 in 100 million to 1 in 100 thousand.
[-] You can only buy as many tickets as you can afford.
[-] When you win a ticket[s], it shows you how many tickets     you have played so far.
[-] Ticket Price increased from £10 to £100

===============
Infinite Money Exploit
===============
1. Loan £1,000,000,000
2. Buy 10,000,000 lottery tickets
3. Pay off the loan in full, since you can now afford to do    that.
4. Buy lottery tickets for a guaranteed return on investment.
===============

Section 3: Jobs
We have added even MORE features to jobs in order to win.
[-] Pensions
   [-] When you work for a company for 20 years, you can take 
   a pension. You don't work and earn 50% of your wages.
[-] Promotions
   [-] Every 5 years you get promoted if you can be.

Section 4: Save Game
[-] Proper working savegame :o
   [-] Uses .ool2 save format now
[-] You can LOAD the game!
===============
Notes for old saves [Alpha 11 and older]

These old saves [with a .oll file extension] DO NOT LOAD. They probably will not have backwards-compatibility due to a lack of saved variables.
You may be able to reverse-engineer a save file, but it isn't worth it since the game has very few working features.

I may add a utility to port the save files in future, but that will be way in the future and also quite unlikely.

===============
Notes for Alpha 12 Pre-Releases 1-3

These releases may use the .oll2 file extension used by OpenLife lives, but their saves are incompatible with this release, due to using a different save system. 
===============

Section 5: Anticheat Upgrades
===============
The anticheat has become even more powerful. You can no longer do anything, even if you reach the console. This is because ExitError is raised when you exit the game. This wipes every single user-created object from the game, including variables and functions. This means that every single line of game code gets destroyed, making it impossible to edit values without manually rewriting the code into the interpreter.

Section x: General Additions
===============
[-] Custom forenames/surnames
   [-] Create male.dat for male forenames
   [-] female.dat for female forenames
   [-] surname.dat for surnames
   [-] Each file can contain a list of custom surnames,
       each split by a new line.
   [-] This allows you to enhance the game experience and 
       have more variety with names.
[-] In the main menu, there is a new option to download the     latest character. It is currently unused.
   [-] You will be able to download a .oll2 file from the        WinFan3672 FTP server and have a custom life to               use in OL.This will be implemented later.
[-] Loans
   [-] 20% interest
      [-] Automatically pays off 10% per year
      [-] As a result,your debt increases forever unless you           pay it off in full. 
   [-] Can pay off loan at once
   [-] Used for:
      [-] Student Loan
      [-] Loan
   [-] No Limit
[-] Savings Account
   [-] Withdraw / Deposit
   [-] 5% per year
   [-] Compound interest
   [-] No limit
[-] Vacation
   [-] Increases happiness
   [-] Costs money
[-] Fertility
   [-] Sperm Donation
      [-] Pay £50 per year
[-] Random Stat Changes
   [-] Stats increase between -10% and +10% each year
   [-] I was an idiot when developing this and forgot to        divide by 100, resulting in either max stats or death.

Section x: Minor Tweaks/Bug Fixes
[-] The game now SELF DESTRUCTS when it exits, meaning you cannot cheat anymore :)
[-] Job #28 requires 80%+ looks to get [as originally     intended]
[-] Fixed pretty much all bugs where things like pensions would carry over to the next life if you exit to the main menu and made a new life
[-] All currency errors showing dollars are GONE. The pound is now consistent.
[-] When you bust in Blackjack, all that goes into a loan.     That's right, gambling debt.
[-] Going to the gym increases your happiness by 45%, up from     25%.

Section x: Bugs
===============
[-] A lot of variables are not saved in the OLL2 save file.
   [-] This will be rectified in A13.
   [-] The variables [that I know of] are:
      [-] mother_rel
      [-] father_rel
      [-] did_tutor
   [-] These will not be fixed right now for compatibility        reasons, since Alpha 12 PR3 save files would be        broken, and   also because I would have a lot of              broken save files.
[-] Not all of the job roles have a text, instead saying     "Corporate Slave".

Section x: Suspected bugs
===============
[-] The anticheat is bypassable, somehow. I just know it.
[-] The promotions system definitely has issues that I don't     know about. Report them, please.
===============
Build Date  : 7 Oct 2022
Release Date: 7 Oct 2022
Next Update : The Childbirth Update