Alpha 13
Alpha 13- The Childbirth Update

1 Introduction
1.1 What This Update Is

Alpha 13 is here! It features a lot of great features that you have been dying to have added, such as:
[-] Spouses

This update is also the first time that you can change the currency from in-game, by editing the main file!
Simply find the currency variable near the top and set it to whatever you would like, such as "$" or "♪"!

It was originally going to have children, but that was canned due to complexity. Children will [obviously] be added, it's just that it's on the back burner for now.

1.2 Alpha 14

Alpha 14 will be the identity update. It will bring things like fame and friends. Be excited.

1.3 Thoughts

This update feels kind of half-arsed, compared to Alpha 12. I think this update is a minor let-down, but then again, I am taking the progress of the game for granted here. This update is smaller and took longer to make, but that's fine. I just need to remind myself that.

2 Spouses
2.1 Spouse

You can have a spouse and:
[-] Marry them
[-] Break up with them
Sadly, since it's an alpha, it will always be a female spouse and you will always be a male character.

3 Die()
3.1 What Is It
The die() function is a universal death screen. It has 5 causes at the moment, and not all of them in use [Suicide, heart attack, health issues, cancer and old age].

3.2 Menu
The menu will be the front page of the OpenLife Times, a newspaper in the OpenLife universe. You will be able to read the article, and once you've read it you can:
[-] Exit the game
[-] Return to the main menu
[-] Respawn
   [-] That's right, escape death! It only works up to 130 years old.

4 Job Additions
4.1 Teen Jobs

You can now have a job as a teen!
The jobs are:
[-] Pizza Delivery Person
[-] Newspaper Delivery Person
[-] Lawn Mower
[-] Car Washer

4.2 Part-Time Jobs

You can now have a job alongside your job :o
You work 10 hours per week, so 520 hours. 
The jobs are:
[-] Receptionist
[-] Mall Santa

5 School Gang
5.1 Gang
This is a half-arsed update to school gangs. It adds:
[-] Gang wars
[-] Enemy gangs

The enemy gang has a placeholder name, sadly.

6 General Additions
6.1 Currency Changes
The variable `currency` determines what symbol is rendered when you see the `£` symbol. You can edit the main file and turn it into dollars if you want.
6.2 Custom Surnames
Yep. Make a file called `surname.dat` and every line is a different valid surname. 
6.3 SGCU
SGCU [Save Game Conversion Utility] can convert an Alpha 12 save file into an Alpha 13 one. One day, it will support the original OLL files.
6.4 Options Menu
The options menu allows you to change themes and enable Auto Deposit for the first time ever. No more editing save files. Ok, technically you still will, since the changes get reset when you relaunch the game/exit to the main menu. But still, it's a step in the right direction.

7 Bug Fixes
[-] Fixed a bug where 2 `div()` elements are displayed when rendering the `Feature Not In Current Build` message on the main menu.

8 Known Bugs
[-] Save game breaks a few things

Next Update: Identity Update
Build Date: 26 Oct 2022
Publish Date: 26 Oct 2022