Alpha 16
Alpha 16- The "Bugfix" Update
====================
Alpha 16 is a rushed update. I feel that if I do not finish it, I will automatically not want to work on the game. 
====================
Additions
====================
[-] Lottery wins
   [-] A value that counts how many times you've won the lottery
[-] New Ribbon- Cheater
   [-] Win the lottery 50+ times
[-] Updated relations menu
   [-] Child list
      [-] A less cluttered way of viewing all of your children.
====================
Bug Fixes
====================
[-] Fixed a double div() element when exiting the game without saving it. 

====================
Delay
====================
Alpha 16 was delayed by about half a week because of SGCU issues. It had trouble converting from Alpha 14 to Alpha 15.